+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Non-Backward Compatible Changes
 - DagNode.isVisible:  has a new flag, checkOverride, which is on by default, and considers visibility override settings
 - referenceQuery/FileReference.getReferenceEdits: if only one of successfulEdits/failedEdits is given, and it is false, we now assume that the desire is to return the other type (and set that flag to true); formerly, this would result in NO edits being returned
 - parent no longer raises an error if setting an object's parent to it's current parent; this makes it behave similarly to the mel command, and to DagNode.setParent
+- renameFile now automatically sets the 'type' if none is supplied (helps avoid renaming a file to 'foo.ma', then saving it as 'mayaBinary')
 
 ----------------------------------
 Changes